업소 상세 조회 api 구현

+++ src/main/java/com/example/alddeul_babsang/service/MapService.java
@@ -3,6 +3,7 @@ package com.example.alddeul_babsang.service;
 import com.example.alddeul_babsang.apiPayload.code.status.ErrorStatus;
 import com.example.alddeul_babsang.apiPayload.exception.handler.TempHandler;
 import com.example.alddeul_babsang.converter.StoreConverter;
+import com.example.alddeul_babsang.entity.Menu;
 import com.example.alddeul_babsang.entity.Store;
 import com.example.alddeul_babsang.repository.StoreRepository;
 import com.example.alddeul_babsang.web.dto.StoreDTO;
@@ -18,7 +19,7 @@ public class MapService {
 
     private final StoreRepository storeRepository;
 
-    // 지도용 착한 업소 리스트 조회
+    // 착한 업소 리스트 조회
     public List<StoreDTO.MapStore> getMapStoreList() {
         List<Store> stores = storeRepository.findAll();
         // 예비밥상 조회
@@ -28,8 +29,8 @@ public class MapService {
                 .collect(Collectors.toList());
     }
 
-    // 특정 착한 업소 조회
-    public StoreDTO.StoreInfo getStore(int storeId) {
+    // 착한 업소 조회
+    public StoreDTO.StoreInfo getStore(Long storeId) {
         // store id 조회 -> 예외 처리
         Store store = storeRepository.findById(storeId)
                 .orElseThrow(() -> new TempHandler(ErrorStatus.STORE_ERROR_ID));
